modularised the code. tables now created from own files
